third_party/WebKit/LayoutTests/platform/win/virtual/mojo-loading/http/tests/budget/reserve-expected.txt

This is a testharness.js-based test.
PASS BudgetService.Reserve should return correct results. 
PASS BudgetService.Reserve should return Type Error if an invalid argument is provided. 
Harness: the test ran to completion.
